Clean up mis-labeled Agreement test, add Policy test
