+++ test/gherkin/tests/test_get_policy.py
@@ -11,27 +11,27 @@ scenarios("get_policy.feature")
 
 
 @given(
-    "an Agreement for MCC Registration exists",
-    target_fixture="mcc_registration_create"
+    "I have an Agreement for MCC Registration",
+    target_fixture="mcc_registration_agreement"
 )
-def mcc_registration_create():
-    # Currently, this is just mocked in Caddy, so we don't create the object,
-    # we just return a hard-coded ID
-    return 1
-
+def mcc_i_have_an_agreement(api_url, client):
+    agreement = client.get(
+        os.path.join(api_url, f"service/agreement/1/")
+    )
+    return json.loads(agreement.content)
 
 @when(
-    "I fetch an Agreement for MCC Registration",
-    target_fixture="api_mcc_registration"
+    "I fetch the Policy of the MCC Registration Agreement",
+    target_fixture="api_mcc_registration_policy"
 )
-def when_api_basic_public_call(api_url, mcc_registration_create, client):
+def when_api_policy_call(api_url, mcc_registration_agreement, client):
     return client.get(
-        os.path.join(api_url, f"service/agreement/{mcc_registration_create}/")
+        os.path.join(api_url, "service/policy/{}/".format(mcc_registration_agreement["id"]))
     )
 
 
-@then("I get a valid Agreement")
-def json_data_is_valid(api_mcc_registration):
-    assert api_mcc_registration.status_code == 200
-    json_data = json.loads(api_mcc_registration.content)
-    assert json_data["id"] == "1"
+@then("I get a valid Policy")
+def json_data_is_valid(api_mcc_registration_policy):
+    assert api_mcc_registration_policy.status_code == 200
+    json_data = json.loads(api_mcc_registration_policy.content)
+    assert json_data["id"] == 1
